add new file fsnparams.go to put fsn call params separately
